% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sorting_steps_matrix.R
\name{as.SortingStepsMatrix}
\alias{as.SortingStepsMatrix}
\title{Convert to SortingStepsMatrix}
\usage{
as.SortingStepsMatrix(x)
}
\arguments{
\item{x}{A matrix representing sorting steps.}
}
\value{
A matrix of class 'SortingStepsMatrix'.
}
\description{
This function takes a matrix and converts it into a SortingStepsMatrix object,
which is a matrix with an additional class attribute used for specialized
sorting step analysis and visualization in the package.
}
