% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sorting_animation.R
\name{create_sorting_animation.SortingStepsMatrix}
\alias{create_sorting_animation.SortingStepsMatrix}
\title{Create Sorting Animation}
\usage{
\method{create_sorting_animation}{SortingStepsMatrix}(sorting_steps_df)
}
\arguments{
\item{sorting_steps_df}{A data frame or matrix where each row represents the state of the sequence at a particular step of sorting.
The columns represent the elements in the sequence.}
}
\value{
An animation object that visually represents the sorting process step-by-step.
Each frame of the animation corresponds to one step in the sorting process.
}
\description{
Generates an animated visualization of a sorting process, using the steps recorded during the sorting.
}
\examples{
#Not run because the resulting gif comes out as multiple .pngs which gives a note.
#create_sorting_animation(bubble_sort_steps(10))

}
