% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sorting_animation.R
\name{find_differences_vectorized}
\alias{find_differences_vectorized}
\title{This takes in a matrix, checks for the differences between the rows in the matrix}
\usage{
find_differences_vectorized(matrix)
}
\arguments{
\item{matrix}{the matrix of sorting steps}
}
\value{
list of differences between the rows
}
\description{
This takes in a matrix, checks for the differences between the rows in the matrix
}
