% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertion_sort_steps.R
\name{insertion_sort_steps}
\alias{insertion_sort_steps}
\title{Takes positive integer N, creates random vector which is first step in the sort, appends each step to a matrix}
\usage{
insertion_sort_steps(N)
}
\arguments{
\item{N}{positive integer}
}
\value{
matrix of steps
}
\description{
Takes positive integer N, creates random vector which is first step in the sort, appends each step to a matrix
}
\examples{
# example code
insertion_sort_steps(10)
}
