% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sorting_steps_matrix.R
\name{as.SortingStepsMatrix}
\alias{as.SortingStepsMatrix}
\title{Convert Matrix to SortingStepsMatrix Object}
\usage{
as.SortingStepsMatrix(x)
}
\arguments{
\item{x}{A matrix representing sorting steps.}
}
\value{
A SortingStepsMatrix object that represents the input matrix as sorting steps.
}
\description{
This function converts an input matrix into a SortingStepsMatrix object, which is used to represent sorting steps.
}
\examples{
steps_matrix <- matrix(1:9, ncol = 3)
sorted_steps <- as.SortingStepsMatrix(steps_matrix)
}
