% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sorting_steps_matrix.R
\name{as.SortingStepsMatrix}
\alias{as.SortingStepsMatrix}
\title{Convert to SortingStepsMatrix}
\usage{
as.SortingStepsMatrix(x)
}
\arguments{
\item{x}{A matrix representing sorting steps.}
}
\value{
A matrix of class 'SortingStepsMatrix'.
}
\description{
This function takes a matrix and converts it into a SortingStepsMatrix object,
which is a matrix with an additional class attribute used for specialized
sorting step analysis and visualization in the package.
}
\examples{
# Create a vector of numbers
numbers <- 1:9
# Create a 3x3 matrix from the vector
my_matrix <- matrix(numbers, nrow = 3, ncol = 3)
as.SortingStepsMatrix(my_matrix)
}
